==> Notebooks/02 - Regular Season Statistics.ipynb <==
## Standard Regular Season Statistics

This notebook is where I created the regular seasson statistics as well as the offensive/defensive efficiency numbers

import pandas as pd

regSeason = pd.read_csv("../raw data/RegularSeasonDetailedResults.csv")

print len(regSeason)
print len(regSeason.columns)

regSeason.head()

print regSeason.columns

#create two data frames: one for winning teams one for losing teams. Also add in the opposing team stats

wTeams = regSeason[['Season','Wteam','Daynum','Wscore','Numot','Wfgm','Wfga','Wfgm3',
                    'Wfga3','Wftm','Wfta','Wor','Wdr','Wast','Wto','Wstl','Wblk','Wpf',
                    'Lscore','Lfgm','Lfga','Lfgm3','Lfga3','Lftm','Lfta','Lor','Ldr',
                    'Last','Lto','Lstl','Lblk','Lpf']]



lTeams = regSeason[['Season','Lteam','Daynum','Lscore','Numot','Lfgm','Lfga','Lfgm3',
                    'Lfga3','Lftm','Lfta','Lor','Ldr','Last','Lto','Lstl','Lblk','Lpf',
                    'Wscore','Wfgm','Wfga','Wfgm3','Wfga3','Wftm','Wfta','Wor','Wdr',
                    'Wast','Wto','Wstl','Wblk','Wpf']]

wTeams['win'] = 1

wTeams.head()

lTeams['win'] = 0

lTeams.head()

wTeams.columns = ['season','team','daynum','score','numot','fgm','fga','fgm3',
                  'fga3','ftm','fta','or','dr','ast','to','stl','blk','pf',
                  'opp_score','opp_fgm','opp_fga','opp_fgm3','opp_fga3','opp_ftm',
                  'opp_fta','opp_or','opp_dr','opp_ast','opp_to','opp_stl',
                  'opp_blk', 'opp_pf', 'win']

lTeams.columns = ['season','team','daynum','score','numot','fgm','fga','fgm3',
                  'fga3','ftm','fta','or','dr','ast','to','stl','blk','pf',
                  'opp_score','opp_fgm','opp_fga','opp_fgm3','opp_fga3','opp_ftm',
                  'opp_fta','opp_or','opp_dr','opp_ast','opp_to','opp_stl',
                  'opp_blk', 'opp_pf', 'win']

regularSeason = wTeams.append(lTeams)

regularSeason.head()

import numpy as np

seasonLongStats = regularSeason.groupby(['season','team'], as_index = False).agg({'fgm':sum, 'fga': sum,
                                                                'fgm3':sum, 'fga3': sum,
                                                                'ftm':sum, 'fta': sum,
                                                                'or': np.mean, 'dr': np.mean,
                                                                'ast': np.mean, 'to': np.mean,
                                                                'stl': np.mean, 'blk': np.mean,
                                                                'pf': np.mean, 'score': np.mean,
                                                                'opp_fgm':sum, 'opp_fga': sum,
                                                                'opp_fgm3':sum, 'opp_fga3': sum,
                                                                'opp_ftm':sum, 'opp_fta': sum,
                                                                'opp_or': np.mean, 'opp_dr': np.mean,
                                                                'opp_ast': np.mean, 'opp_to': np.mean,
                                                                'opp_stl': np.mean, 'opp_blk': np.mean,
                                                                'opp_pf': np.mean, 'opp_score': np.mean})

seasonLongStats.head()

seasonLongStats['fgp'] = seasonLongStats['fgm'] / seasonLongStats['fga']
seasonLongStats['fg3p'] = seasonLongStats['fgm3'] / seasonLongStats['fga3']
seasonLongStats['ftp'] = seasonLongStats['ftm'] / seasonLongStats['fta']

seasonLongStats['opp_fgp'] = seasonLongStats['opp_fgm'] / seasonLongStats['opp_fga']
seasonLongStats['opp_fg3p'] = seasonLongStats['opp_fgm3'] / seasonLongStats['opp_fga3']
seasonLongStats['opp_ftp'] = seasonLongStats['opp_ftm'] / seasonLongStats['opp_fta']

seasonLongStats.drop(['fgm', 'fga', 'fgm3', 'fga3', 'ftm', 'fta',
                     'opp_fgm', 'opp_fga', 'opp_fgm3', 'opp_fga3', 'opp_ftm', 'opp_fta'], axis = 1, inplace = True)

print seasonLongStats.columns

seasonLongStats.columns = ['season', 'team', 'opp_toAvg', 'opp_scoreAvg', 'opp_orAvg', 'toAvg', 'opp_stlAvg',
                           'scoreAvg', 'pfAvg', 'blkAvg', 'astAvg', 'opp_astAvg', 'opp_drAvg', 'drAvg', 'opp_blkAvg',
                           'stlAvg', 'opp_pfAvg', 'orAvg', 'fgp', 'fg3p', 'ftp', 'opp_fgp', 'opp_fg3p', 'opp_ftp']

seasonLongStats.head()

## Offensive Ratings

100 x Pts / 0.5 * ((Tm FGA + 0.4 * Tm FTA - 1.07 * (Tm ORB / (Tm ORB + Opp DRB)) * (Tm FGA - Tm FG) + Tm TOV) + (Opp FGA + 0.4 * Opp FTA - 1.07 * (Opp ORB / (Opp ORB + Tm DRB)) * (Opp FGA - Opp FG) + Opp TOV))[1]

print regularSeason.columns

efficiencyStats = regularSeason.groupby(['season','team'], as_index = False).agg({'score':sum, 'fga': sum,
                                                                                  'fta':sum, 'or': sum,
                                                                                  'opp_dr': sum, 'fgm': sum,
                                                                                  'to': sum,
                                                                                  'opp_fga': sum, 'opp_fta': sum,
                                                                                  'opp_or': sum, 'dr': sum,
                                                                                  'opp_fgm':sum, 'opp_to': sum,
                                                                                  'opp_score': sum})

efficiencyStats['OffRating'] = (100. * efficiencyStats['score']) / (0.5 * ((efficiencyStats['fga'] + 0.4 * efficiencyStats['fta'] - 1.07 * (efficiencyStats['or'] / (efficiencyStats['or'] + efficiencyStats['opp_dr'])) * (efficiencyStats['fga'] - efficiencyStats['fgm']) + efficiencyStats['to']) + (efficiencyStats['opp_fga'] + 0.4 * efficiencyStats['opp_fta'] - 1.07 * (efficiencyStats['opp_or'] / (efficiencyStats['opp_or'] + efficiencyStats['dr'])) * (efficiencyStats['opp_fga'] - efficiencyStats['opp_fgm']) + efficiencyStats['opp_to'])))





efficiencyStats['DefRating'] = (efficiencyStats['opp_score'] / (efficiencyStats['opp_fga'] + 0.475 * efficiencyStats['opp_fta'] - efficiencyStats['opp_or'] + efficiencyStats['opp_to']))*100.




efficiencyStats.tail()

seasonLongStats = pd.merge(seasonLongStats, efficiencyStats[['season','team', 'OffRating','DefRating']],  how='left',
              left_on=['season','team'], right_on = ['season','team'])

seasonLongStats.head()

seasonLongStats.to_csv("../derived data/seasonLongStats.csv")
