==> Notebooks/03 - ELO Creation.ipynb <==
## ELO Creation

This notebook is where I calculated ELO rating for each team

## Import and clean data

import pandas as pd
import numpy as np
from sklearn.preprocessing import LabelEncoder

regSeason = pd.read_csv("../raw data/RegularSeasonCompactResults.csv")
tournament = pd.read_csv("../raw data/TourneyCompactResults.csv")

# ELO constants
mean_elo = 1500.
elo_width = 400.
k_factor = 64.

df = pd.concat((regSeason, tournament), ignore_index= True)

df.columns

df.drop(['Wscore', 'Lscore', 'Wloc', 'Numot'], inplace= True, axis = 1)

df.sort_values(by=['Season', 'Daynum'], inplace=True)

df.head()

## Create ELO Functions

def update_elo(winner_elo, loser_elo):

    expected_win = expected_result(winner_elo, loser_elo)
    change_in_elo = k_factor * (1-expected_win)
    winner_elo += change_in_elo
    loser_elo -= change_in_elo
    return winner_elo, loser_elo

def expected_result(elo_a, elo_b):

    expect_a = 1.0/(1+10**((elo_b - elo_a)/elo_width))
    return expect_a

def update_end_of_season(elos):

    diff_from_mean = elos - mean_elo
    elos -= diff_from_mean/3.0
    return elos

df['w_elo_before_game'] = 0
df['w_elo_after_game'] = 0
df['l_elo_before_game'] = 0
df['l_elo_after_game'] = 0
elo_per_season = {}
n_teams = len(np.unique(df[['Wteam', 'Lteam']]))
current_elos = np.ones(shape=(n_teams)) * mean_elo

## Create a unique day count

df['total_days'] = (df.Season-1980)*365.25 + df.Daynum

df.head()

df_team_elos = pd.DataFrame(index=df.total_days.unique(), 
                            #columns=range(n_teams))
                            columns = np.unique(df[['Wteam', 'Lteam']]))
df_team_elos.iloc[0, :] = current_elos

df_team_elos.head()

## ELO Loop time

* We go through each row in the DataFrame.
* We look up the current Elo rating of both teams.
* We calculate the expected wins for the team that actually won. This is also what we use for probability of winning.
* Write Elo before and after the game in the Data Frame.
* Update the Elo rating for both teams in the "current_elos" list.

current_season = df.at[0, 'Season']
for row in df.itertuples():
    if row.Season != current_season:
        # Check if we are starting a new season. 
        # Regress all ratings towards the mean
        current_elos = update_end_of_season(current_elos)
        # Write the beginning of new season ratings to a dict for later lookups.
        elo_per_season[row.Season] = current_elos.copy()
        current_season = row.Season
    idx = row.Index
    w_id = row.Wteam
    l_id = row.Lteam
    # Get current elos
    #w_elo_before = current_elos[w_id]
    w_elo_before = current_elos[df_team_elos.columns.get_loc(w_id)]
    #l_elo_before = current_elos[l_id]
    l_elo_before = current_elos[df_team_elos.columns.get_loc(l_id)]
    # Update on game results
    w_elo_after, l_elo_after = update_elo(w_elo_before, l_elo_before)
        
    # Save updated elos
    df.at[idx, 'w_elo_before_game'] = w_elo_before
    df.at[idx, 'l_elo_before_game'] = l_elo_before
    df.at[idx, 'w_elo_after_game'] = w_elo_after
    df.at[idx, 'l_elo_after_game'] = l_elo_after
    current_elos[df_team_elos.columns.get_loc(w_id)] = w_elo_after
    current_elos[df_team_elos.columns.get_loc(l_id)] = l_elo_after
    
    # Save elos to team DataFrame
    today = row.total_days
    df_team_elos.at[today, w_id] = w_elo_after
    df_team_elos.at[today, l_id] = l_elo_after

df.head()

n_samples = 10000
samples = df[df.Season > 2010].sample(n_samples)
loss=0
expected_list = []
for row in samples.itertuples():
    w_elo = row.w_elo_before_game
    l_elo = row.l_elo_before_game
    w_expected = expected_result(w_elo, l_elo)
    expected_list.append(w_expected)
    loss += np.log(w_expected)
print(loss/n_samples)

import matplotlib.pyplot as plt
import seaborn as sns

%matplotlib notebook

sns.distplot(expected_list, kde=False, bins=20)
plt.xlabel('Elo Expected Wins for Actual Winner')
plt.ylabel('Counts')

df_team_elos.head()

df.to_csv("../derived data/eloGames.csv")
