==> Notebooks/04 - Modeling.ipynb <==
## Modeling

In this notebook I combined the season long stats and ELO rankings from notebooks 02 and 03 into the initial NCAA data frame from kaggle and evaluated some models. The results secion at the end is where I evaluated the 2016 games for each team and then put it into the kaggle submission format to help me create my brackets

import pandas as pd
import numpy as np

regSeason = pd.read_csv("../raw data/RegularSeasonDetailedResults.csv")
tourney = pd.read_csv("../raw data/TourneyDetailedResults.csv")
tourneySeeds = pd.read_csv("../raw data/TourneySeeds.csv")
tourneySlots = pd.read_csv("../raw data/TourneySlots.csv")

eloStats = pd.read_csv("../derived data/eloGames.csv")
seasonStats = pd.read_csv("../derived data/seasonLongStats.csv")

eloStats.drop(eloStats.columns[0], axis=1, inplace=True)
seasonStats.drop(seasonStats.columns[0], axis=1, inplace=True)

#create two data frames: one for winning teams one for losing teams. Also add in the opposing team stats

wTeams = regSeason[['Season','Wteam','Lteam','Daynum','Wscore','Numot','Wfgm','Wfga','Wfgm3',
                    'Wfga3','Wftm','Wfta','Wor','Wdr','Wast','Wto','Wstl','Wblk','Wpf',
                    'Lscore','Lfgm','Lfga','Lfgm3','Lfga3','Lftm','Lfta','Lor','Ldr',
                    'Last','Lto','Lstl','Lblk','Lpf']]



lTeams = regSeason[['Season','Lteam','Wteam','Daynum','Lscore','Numot','Lfgm','Lfga','Lfgm3',
                    'Lfga3','Lftm','Lfta','Lor','Ldr','Last','Lto','Lstl','Lblk','Lpf',
                    'Wscore','Wfgm','Wfga','Wfgm3','Wfga3','Wftm','Wfta','Wor','Wdr',
                    'Wast','Wto','Wstl','Wblk','Wpf']]

wTeams['win'] = 1
lTeams['win'] = 0

wTeams.columns = ['season','team','opp','daynum','score','numot','fgm','fga','fgm3',
                  'fga3','ftm','fta','or','dr','ast','to','stl','blk','pf',
                  'opp_score','opp_fgm','opp_fga','opp_fgm3','opp_fga3','opp_ftm',
                  'opp_fta','opp_or','opp_dr','opp_ast','opp_to','opp_stl',
                  'opp_blk', 'opp_pf', 'win']

lTeams.columns = ['season','team','opp','daynum','score','numot','fgm','fga','fgm3',
                  'fga3','ftm','fta','or','dr','ast','to','stl','blk','pf',
                  'opp_score','opp_fgm','opp_fga','opp_fgm3','opp_fga3','opp_ftm',
                  'opp_fta','opp_or','opp_dr','opp_ast','opp_to','opp_stl',
                  'opp_blk', 'opp_pf', 'win']

regularSeason = wTeams.append(lTeams)

print regularSeason.columns

regularSeason = regularSeason[['season','daynum','team','opp','win']]

print(seasonStats.columns)

df = pd.merge(regularSeason, seasonStats[['season', 'team', 'toAvg','scoreAvg', 'pfAvg', 'blkAvg', 'astAvg', 'drAvg',
'stlAvg', 'orAvg', 'fgp', 'fg3p', 'ftp', 'OffRating', 'DefRating']],  how='left',
              left_on=['season','team'], right_on = ['season','team'])

df = pd.merge(df, seasonStats[['season', 'team', 'toAvg','scoreAvg', 'pfAvg', 'blkAvg', 'astAvg', 'drAvg',
'stlAvg', 'orAvg', 'fgp', 'fg3p', 'ftp', 'OffRating', 'DefRating']],  how='left',
              left_on=['season','opp'], right_on = ['season','team'])

print df.columns

df['toDif'] = df['toAvg_x'] - df['toAvg_y']
df['scoreDif'] = df['scoreAvg_x'] - df['scoreAvg_y']
df['pfDif'] = df['pfAvg_x'] - df['pfAvg_y']
df['blkDif'] = df['blkAvg_x'] - df['blkAvg_y']
df['astDif'] = df['astAvg_x'] - df['astAvg_y']
df['drDif'] = df['drAvg_x'] - df['drAvg_y']
df['stlDif'] = df['stlAvg_x'] - df['stlAvg_y']
df['orDif'] = df['orAvg_x'] - df['orAvg_y']
df['fgpDif'] = df['fgp_x'] - df['fgp_y']
df['fg3pDif'] = df['fg3p_x'] - df['fg3p_y']
df['ftpDif'] = df['ftp_x'] - df['ftp_y']
df['offRatDif'] = df['OffRating_x'] - df['OffRating_y']
df['defRatDif'] = df['DefRating_x'] - df['DefRating_y']

df.drop(['toAvg_x',
       'scoreAvg_x', 'pfAvg_x', 'blkAvg_x', 'astAvg_x', 'drAvg_x',
       'stlAvg_x', 'orAvg_x', 'fgp_x', 'fg3p_x', 'ftp_x', 'OffRating_x',
       'DefRating_x', 'team_y', 'toAvg_y', 'scoreAvg_y', 'pfAvg_y',
       'blkAvg_y', 'astAvg_y', 'drAvg_y', 'stlAvg_y', 'orAvg_y', 'fgp_y',
       'fg3p_y', 'ftp_y', 'OffRating_y', 'DefRating_y'], axis = 1, inplace = True)

df.head()

df.rename(columns={'team_x':'team'}, inplace=True)

eloStats.head()

eloWin = eloStats[['Season', 'Daynum', 'Wteam', 'w_elo_before_game']]
eloLose = eloStats[['Season', 'Daynum', 'Lteam', 'l_elo_before_game']]

eloWin.columns = ['season','daynum','team','elo']
eloLose.columns = ['season','daynum','team','elo']

eloTotals = eloWin.append(eloLose)

df = pd.merge(df, eloTotals,  how='left',
              left_on=['season','daynum','team'], right_on = ['season','daynum','team'])

df = pd.merge(df, eloTotals,  how='left',
              left_on=['season','daynum','opp'], right_on = ['season','daynum','team'])

print df.columns

df.drop('team_y', axis = 1, inplace = True)

df['eloDif'] = df['elo_x'] - df['elo_y']

print df.columns

## MODELING TIME!

import sklearn.model_selection as ms

X_train, X_test, y_train, y_test = ms.train_test_split(df[['toDif', 'scoreDif', 'pfDif', 'blkDif', 'astDif',
                                                           'drDif', 'stlDif', 'orDif', 'fgpDif', 'fg3pDif',
                                                           'ftpDif', 'offRatDif', 'defRatDif','eloDif']]
                                                       , df['win'], test_size=0.2, random_state=42)

## Logistic Regression

#Start with a vanilla regressor
import sklearn.linear_model as lm
from sklearn.linear_model import LogisticRegression

lr = ms.GridSearchCV(estimator=LogisticRegression(class_weight = 'balanced', penalty = 'l1'),
                                  param_grid={'C': range(11,15,1)},
                                  scoring='neg_log_loss'
)

lr.fit(X_train, y_train)

print(lr.best_score_)
print(lr.best_params_)

Standard Logistic regression can get me to 0.52337282409

from sklearn.metrics import log_loss

predicted = lr.predict_proba(X_test)[:, 1]
log_loss(y_test, predicted)

## Lasso Path

import matplotlib.pyplot as plt
import sklearn.linear_model as lm

%matplotlib notebook
alpha, coef, _ = lm.lasso_path(X_train, y_train)
for i in range(14):
    plt.plot(np.log(alpha), coef[i, :], label=X_train.columns[i])
plt.legend(loc='lower right')

## XGBoost

import xgboost as xgb
from xgboost.sklearn import XGBClassifier

#xgbmodel = ms.GridSearchCV(XGBClassifier(n_estimators=58,max_depth=3, min_child_weight=2,
#                                        gamma = 0, colsample_bytree=0.8),
#                           param_grid={'subsample': np.arange(0.8, 1, 0.1)},
#                           scoring='neg_log_loss',
#                           n_jobs = -1,
#                          verbose = 3)

xgbmodel = XGBClassifier(n_estimators=58,max_depth=3, min_child_weight=2, gamma = 0,
                        subsample=0.8, colsample_bytree=0.8)

xgbmodel.fit(X_train, y_train)

print(xgbmodel.best_score_)

print(xgbmodel.best_params_)

XGBoost looks like it can get me to -0.524869145553

xgpredicted = xgbmodel.predict_proba(X_test)[:, 1]
log_loss(y_test, xgpredicted)

# plot feature importance
def get_xgb_imp(xgb, feat_names):
    from numpy import array
    imp_vals = xgb.booster().get_fscore()
    imp_dict = {feat_names[i]:float(imp_vals.get('f'+str(i),0.)) for i in range(len(feat_names))}
    total = array(imp_dict.values()).sum()
    return {k:v/total for k,v in imp_dict.items()}

get_xgb_imp(xgbmodel,X_train.columns)

## Descision Tree

from sklearn.tree import DecisionTreeClassifier

#treeModel = ms.GridSearchCV(DecisionTreeClassifier(),
#                            {'max_depth': range(2, 10, 1),
#                             'min_samples_leaf': range(5, 25, 5)},
#                            scoring='neg_log_loss',
#                            verbose = 3,
#                            n_jobs = -1)

treeModel = DecisionTreeClassifier(max_depth= 6, min_samples_leaf= 20)

treeModel.fit(X_train, y_train)

print(treeModel.best_score_)
print(treeModel.best_params_)

A decision tree can get to 0.534390103677

DTpredicted = treeModel.predict_proba(X_test)[:, 1]
log_loss(y_test, DTpredicted)

## Random Forest

from sklearn.ensemble import RandomForestClassifier

forestModel=ms.GridSearchCV(RandomForestClassifier(),
                     {'n_estimators': range(150, 201, 25)},#,
                      #'max_depth': range(2, 20, 2),
                      #'min_samples_leaf': range(10, 25, 5),
                      #'max_features': range(2, 4)},
                     scoring='neg_log_loss',
                           verbose = 3,
                           n_jobs = -1)
forestModel.fit(X_train, y_train)

print(forestModel.best_score_)
print(forestModel.best_params_)

RFpredicted = forestModel.predict_proba(X_test)[:, 1]
log_loss(y_test, RFpredicted)

## RESULTS

These are the results for all possible 2016 combinations. At the end I put them into the official Kaggle submission format so I could use a script I found to output tournament results

import itertools
games =  pd.DataFrame(list(itertools.combinations(tourneySeeds.loc[(tourneySeeds['Season'] == 2016)]['Team'], 2)))

games['season'] = 2016

games.columns = ['team','opp','season']

games = pd.merge(games, seasonStats[['season', 'team', 'toAvg','scoreAvg', 'pfAvg', 'blkAvg', 'astAvg', 'drAvg',
'stlAvg', 'orAvg', 'fgp', 'fg3p', 'ftp', 'OffRating', 'DefRating']],  how='left',
              left_on=['season','team'], right_on = ['season','team'])

games = pd.merge(games, seasonStats[['season', 'team', 'toAvg','scoreAvg', 'pfAvg', 'blkAvg', 'astAvg', 'drAvg',
'stlAvg', 'orAvg', 'fgp', 'fg3p', 'ftp', 'OffRating', 'DefRating']],  how='left',
              left_on=['season','opp'], right_on = ['season','team'])

games['toDif'] = games['toAvg_x'] - games['toAvg_y']
games['scoreDif'] = games['scoreAvg_x'] - games['scoreAvg_y']
games['pfDif'] = games['pfAvg_x'] - games['pfAvg_y']
games['blkDif'] = games['blkAvg_x'] - games['blkAvg_y']
games['astDif'] = games['astAvg_x'] - games['astAvg_y']
games['drDif'] = games['drAvg_x'] - games['drAvg_y']
games['stlDif'] = games['stlAvg_x'] - games['stlAvg_y']
games['orDif'] = games['orAvg_x'] - games['orAvg_y']
games['fgpDif'] = games['fgp_x'] - games['fgp_y']
games['fg3pDif'] = games['fg3p_x'] - games['fg3p_y']
games['ftpDif'] = games['ftp_x'] - games['ftp_y']
games['offRatDif'] = games['OffRating_x'] - games['OffRating_y']
games['defRatDif'] = games['DefRating_x'] - games['DefRating_y']

games.drop(['toAvg_x',
       'scoreAvg_x', 'pfAvg_x', 'blkAvg_x', 'astAvg_x', 'drAvg_x',
       'stlAvg_x', 'orAvg_x', 'fgp_x', 'fg3p_x', 'ftp_x', 'OffRating_x',
       'DefRating_x', 'team_y', 'toAvg_y', 'scoreAvg_y', 'pfAvg_y',
       'blkAvg_y', 'astAvg_y', 'drAvg_y', 'stlAvg_y', 'orAvg_y', 'fgp_y',
       'fg3p_y', 'ftp_y', 'OffRating_y', 'DefRating_y'], axis = 1, inplace = True)

games.rename(columns={'team_x':'team'}, inplace=True)

games.head()

finalELO = eloTotals.groupby(['season','team'], as_index = False).agg({'daynum': max})

finalELO = pd.merge(finalELO, eloTotals, how='left', left_on=['season','daynum','team'],
                 right_on=['season','daynum','team'])

finalELO.head()

games = pd.merge(games, finalELO,  how='left',
              left_on=['season','team'], right_on = ['season','team'])

games = pd.merge(games, finalELO,  how='left',
              left_on=['season','opp'], right_on = ['season','team'])

games['eloDif'] = games['elo_x'] - games['elo_y']

games.head()

predicted = lr.predict_proba(games[['toDif', 'scoreDif', 'pfDif', 'blkDif', 'astDif','drDif',
                                    'stlDif', 'orDif', 'fgpDif', 'fg3pDif','ftpDif',
                                    'offRatDif', 'defRatDif','eloDif']])[:, 1]  

predicted

games['predicted'] = predicted

games.head()

games.to_csv("../derived data/2016Games.csv")

submissions = games

submissions['Id'] = submissions['season'].astype(str).str.cat(submissions['team_x'].astype(str), sep='_')

submissions.head()

submissions['Id'] = submissions['Id'].astype(str).str.cat(submissions['opp'].astype(str), sep='_')

submissions = submissions[['Id', 'predicted']]

submissions.rename(columns={'predicted':'Pred'}, inplace=True)

submissions.to_csv("../derived data/2016submission.csv", index=False)
